backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pywebpush==2.3.0
python-dotenv==1.0.0
cryptography==41.0.7
jinja2==3.1.2